EOMC: changed some public methods in private one and moved some static functions in eomc_utils.h under namespace yarp::dev::eomc
