+++ src/libraries/icubmod/embObjMotionControl/CMakeLists.txt
@@ -21,7 +21,7 @@ yarp_prepare_plugin(embObjMotionControl
   INCLUDE_DIRECTORIES(${CMAKE_CURRENT_SOURCE_DIR} ${iCubDev_INCLUDE_DIRS})
 
 #    yarp_add_plugin(embObjMotionControl embObjMotionControl.cpp embObjMotionControl.h usrcbk/eOcfg_nvsEP_mc_usrcbk_pippo.c )
-    yarp_add_plugin(embObjMotionControl embObjMotionControl.cpp embObjMotionControl.h mcParser.cpp mcParser.h measuresConverter.cpp measuresConverter.h)
+    yarp_add_plugin(embObjMotionControl embObjMotionControl.cpp embObjMotionControl.h mcParser.cpp mcParser.h measuresConverter.cpp measuresConverter.h eomc_utils.h)
     TARGET_LINK_LIBRARIES(embObjMotionControl ethResources iCubDev ${YARP_LIBRARIES} ${ACE_LIBRARIES})
     icub_export_plugin(embObjMotionControl)
